commit for loose coupling
fixed separation of concerns between NewsServlet and NewsReaderDAO, so
that the DAO-class doesn't touch the ServletRequest.
Changed name of the interface NewsFetcher to NewsReader, and same with
the implementing DAO. Added a primitive prototype of
NewsPublisher-interface with DAO.
Changed name of index.html to api-tester.html.
I put the jdbc-driver in the project(WEB-INF/lib) instead of
tomcat-folder so that the project can(should) start "as is".