I have refactored the newsfeed and newspublish-features
Changes in front-end:
All methods concerning publishing of news have been moved from to the
global newsfeedservice. 
The edit-feature in the newsfeed is now identical to the
publish-feature. The publish-form is now in a partial html, so it can be
included anywhere if needed.

Changes in backend: 
Fixed an issue with published/edited news not being saved in utf-8.
The classes NewsPublisherDAOJDBC and NewsEditorDAOJDBC have been merged
into NewsPublisherDAOJDBC. The editor-class have been removed. The
publishNews()-method will decide to create a new news-entry in the
database, or update an existing one, by checking if the news-entry have
an id. If it doesn't have an id, it's assumed to be a new entry.