Added new features to the SMS-service
Backend: added possibility to check remaining SMS-credits, simple error handling, sending data as JSON in the HTTPServletResponse.
Frontend: It's possible to see the amount of SMS-credits left in the SMS-form. If an error occurs the user will get feedback in form of a message and an optional Error-button that will pop-up a modal describing the error in detail.