Fixed endless loop in newsfeed + changed ip for host
Fixed a bug in newsfeed/controller.js that caused an endless loop if no
important news exists. 

I Also updated the host ip-address in many files: app.js,
dao.properties, META-INF/context.xml. Note: If this project ever gets
revived in future, this should be fixed this the dependency of host-ip
is not not in to many files.